Rename admin controller
